feat(tooltip): move all beyond render to core
