+++ 2nd-gen/packages/core/components/tooltip/Tooltip.base.ts
@@ -9,6 +9,7 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+import { PropertyValues } from 'lit';
 import { property } from 'lit/decorators.js';
 
 import { SpectrumElement } from '@spectrum-web-components/core/element/index.js';
@@ -23,8 +24,12 @@ import {
 /**
  * Abstract base class for the Tooltip component.
  *
- * Declares all public properties and sets `role="tooltip"` on the host element.
- * No rendering logic. No DOM traversal. No Floating UI.
+ * Declares all public properties, sets `role="tooltip"` and `popover="auto"` on
+ * the host element, wires the popover lifecycle (`beforetoggle`, `toggle`,
+ * `transitionend`), dispatches `swc-open`, `swc-close`, `swc-after-open`, and
+ * `swc-after-close` events, resolves the trigger element via `for` or
+ * `triggerElement`, and maintains the `ariaDescribedByElements` relationship on
+ * `open` change. No rendering logic, including placement.
  *
  * @slot - Text label displayed in the tooltip.
  */
@@ -75,7 +80,7 @@ export abstract class TooltipBase extends SpectrumElement {
 
   /**
    * The `id` of the trigger element in the same document tree root.
-   * The SWC layer resolves the trigger via `getRootNode().getElementById(this.for)`.
+   * Resolved via `getRootNode().getElementById(this.for)`.
    * Active from the initial release; drives ARIA relationship wiring on `open` change.
    */
   @property({ attribute: 'for', type: String })
@@ -150,8 +155,112 @@ export abstract class TooltipBase extends SpectrumElement {
   //     IMPLEMENTATION
   // ──────────────────────
 
+  // Reflects the browser's actual popover state. Used to guard showPopover/hidePopover
+  // calls in updated() so the toggle listener can sync this.open without re-triggering
+  // the API (preventing a setter → showPopover → toggle → setter cycle).
+  private get isPopoverOpen(): boolean {
+    return this.matches(':popover-open');
+  }
+
+  private resolveTrigger(): HTMLElement | null {
+    if (this.triggerElement) {
+      return this.triggerElement;
+    }
+    if (this.for) {
+      const root = this.getRootNode() as Document | ShadowRoot;
+      const trigger = root.getElementById(this.for);
+      if (!trigger && window.__swc?.DEBUG) {
+        window.__swc.warn(
+          this,
+          `<${this.localName}> for="${this.for}" did not resolve to an element in the current tree root. Check that the referenced id exists in the same document tree root.`,
+          'https://opensource.adobe.com/spectrum-web-components/components/tooltip/',
+          { level: 'high' }
+        );
+      }
+      return trigger;
+    }
+    return null;
+  }
+
+  private syncAriaRelationship(): void {
+    const trigger = this.resolveTrigger();
+    if (!trigger) {
+      return;
+    }
+    const target = (trigger.shadowRoot?.querySelector('button') ??
+      trigger) as Element & {
+      ariaDescribedByElements: Element[] | null;
+    };
+    const current = target.ariaDescribedByElements ?? [];
+    target.ariaDescribedByElements = this.open
+      ? [...current.filter((el) => el !== this), this]
+      : current.filter((el) => el !== this);
+  }
+
+  private dispatchAfterEvent(isOpen: boolean): void {
+    this.dispatchEvent(
+      new CustomEvent(isOpen ? 'swc-after-open' : 'swc-after-close', {
+        bubbles: true,
+        composed: true,
+      })
+    );
+  }
+
+  private readonly handleBeforeToggle = (event: Event): void => {
+    const { newState } = event as ToggleEvent;
+    const eventName = newState === 'open' ? 'swc-open' : 'swc-close';
+    this.dispatchEvent(
+      new CustomEvent(eventName, { bubbles: true, composed: true })
+    );
+  };
+
+  private readonly handleToggle = (event: Event): void => {
+    const { newState } = event as ToggleEvent;
+    const isOpen = newState === 'open';
+    if (isOpen === this.open) {
+      return;
+    }
+    this.open = isOpen;
+    // When no CSS transition is active, dispatch after-* immediately since transitionend will not fire.
+    if (getComputedStyle(this).transitionDuration === '0s') {
+      this.dispatchAfterEvent(isOpen);
+    }
+  };
+
+  private readonly handleTransitionEnd = (event: TransitionEvent): void => {
+    if (event.target !== this) {
+      return;
+    }
+    this.dispatchAfterEvent(this.open);
+  };
+
+  protected override updated(changedProperties: PropertyValues): void {
+    super.updated(changedProperties);
+    if (changedProperties.has('open')) {
+      if (this.open !== this.isPopoverOpen) {
+        if (this.open) {
+          this.showPopover();
+        } else {
+          this.hidePopover();
+        }
+      }
+      this.syncAriaRelationship();
+    }
+  }
+
   public override connectedCallback(): void {
     super.connectedCallback();
     this.setAttribute('role', 'tooltip');
+    this.setAttribute('popover', 'auto');
+    this.addEventListener('beforetoggle', this.handleBeforeToggle);
+    this.addEventListener('toggle', this.handleToggle);
+    this.addEventListener('transitionend', this.handleTransitionEnd);
+  }
+
+  public override disconnectedCallback(): void {
+    super.disconnectedCallback();
+    this.removeEventListener('beforetoggle', this.handleBeforeToggle);
+    this.removeEventListener('toggle', this.handleToggle);
+    this.removeEventListener('transitionend', this.handleTransitionEnd);
   }
 }
